tools/scene_depth_editor/requirements.txt
numpy>=2.0.0
Pillow>=11.0.0
torch>=2.6.0
transformers>=4.48.0

